+++ packages/js/src/Modules/Verto/webrtc/BaseCall.ts
@@ -124,6 +124,15 @@ export default abstract class BaseCall implements IWebRTCCall {
 
   public extension: string = null;
 
+  /**
+   * Indicates if the peer connection's signaling state has transitioned to 'closed'
+   * while the connection was previously active. Used to determine if the call
+   * can be recovered on reconnection.
+   */
+  get signalingStateClosed(): boolean {
+    return this._signalingStateClosed;
+  }
+
   private _state: State = State.New;
 
   private _prevState: State = State.New;
@@ -153,6 +162,8 @@ export default abstract class BaseCall implements IWebRTCCall {
   > = [];
 
   private _isRemoteDescriptionSet: boolean = false;
+
+  private _signalingStateClosed: boolean = false;
 
   constructor(protected session: BrowserSession, opts?: IVertoCallOptions) {
     const {
@@ -202,6 +213,8 @@ export default abstract class BaseCall implements IWebRTCCall {
     this._onMediaError = this._onMediaError.bind(this);
     this._onPeerConnectionFailureError =
       this._onPeerConnectionFailureError.bind(this);
+    this._onPeerConnectionSignalingStateClosed =
+      this._onPeerConnectionSignalingStateClosed.bind(this);
     this._onTrickleIceSdp = this._onTrickleIceSdp.bind(this);
     this._registerPeerEvents = this._registerPeerEvents.bind(this);
     this._registerTrickleIcePeerEvents =
@@ -444,6 +457,7 @@ export default abstract class BaseCall implements IWebRTCCall {
     this.setState(State.Hangup);
 
     const _close = () => {
+      logger.debug(`[${this.id}] Closing peer from hangup`);
       this.peer?.close();
       return this.setState(State.Destroy);
     };
@@ -1602,6 +1616,17 @@ export default abstract class BaseCall implements IWebRTCCall {
     this.hangup({}, false);
   }
 
+  private _onPeerConnectionSignalingStateClosed(data: any) {
+    this._signalingStateClosed = true;
+    this._dispatchNotification({
+      type: NOTIFICATION_TYPE.signalingStateClosed,
+      ...data,
+    });
+    logger.debug(
+      'Peer connection signaling state closed, call is not recoverable'
+    );
+  }
+
   private _dispatchConferenceUpdate(params: any) {
     this._dispatchNotification({
       type: NOTIFICATION_TYPE.conferenceUpdate,
@@ -1650,6 +1675,11 @@ export default abstract class BaseCall implements IWebRTCCall {
       this._onPeerConnectionFailureError,
       this.id
     );
+    register(
+      SwEvent.PeerConnectionSignalingStateClosed,
+      this._onPeerConnectionSignalingStateClosed,
+      this.id
+    );
     if (isFunction(onNotification)) {
       register(SwEvent.Notification, onNotification.bind(this), this.id);
     }
@@ -1660,12 +1690,14 @@ export default abstract class BaseCall implements IWebRTCCall {
 
   protected _finalize() {
     this._stopStats();
+    logger.debug(`[${this.id}] Closing peer from _finalize`);
     this.peer?.close();
     const { remoteStream, localStream } = this.options;
     stopStream(remoteStream);
     stopStream(localStream);
     deRegister(SwEvent.MediaError, null, this.id);
     deRegister(SwEvent.PeerConnectionFailureError, null, this.id);
+    deRegister(SwEvent.PeerConnectionSignalingStateClosed, null, this.id);
     this.session.calls[this.id] = null;
     delete this.session.calls[this.id];
   }