+++ packages/js/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [2.25.15](https://github.com/team-telnyx/webrtc/compare/webrtc/v2.25.13...webrtc/v2.25.14) (2025-12-16)
+
+- fix: recreate call if peer connection signaling state is closed
+- chore: restart debug stats reporter on session reconnect
+- chore: expose PeerConnectionSignalingStateClosed event
+- chore: update documentation for event and error handling
+
 ## [2.25.14](https://github.com/team-telnyx/webrtc/compare/webrtc/v2.25.13...webrtc/v2.25.14) (2025-12-16)
 
 - fix: make sure offerToReceiveAudio is set if call audio options is undefined